LSShadows

Documentation for LSShadows.

LSShadows.UnitaryCircuit

LSShadows.construct_unitary

LSShadows.onsite_haar_circuit

LSShadows.xzxcnot_circuit

LSShadows.zzx_circuit

LSShadows.UnitaryCircuit — Type

RandomCircuit object. The choice of single-qubit and multi-qubit gates that conform a layer is passed as an argument. Here we support Val(:XZXCnot)

Arguments

N::Int: Number of qubits

p::Int: Circuit depth

LSShadows.construct_unitary — Method

construct_unitary(f::UnitaryCircuit; val=Val(:XZXCnot))

Returns the unitary operator representing the N-qubit and depth p unitary circuit with the choice of layer given by the parameter type

Arguments

f::RandomCircuit: UnitaryCircuit object

Optional arguments

val: By default equal to Val(:XZXCnot). Other options are Val(:ZZX)

corresponding to the Floquet Ising Circuit and Val(:OnsiteHaar) for a local circuit composed of tensor products of Haar-random single qubit unitaries.

LSShadows.onsite_haar_circuit — Method

onsite_haar_circuit(f::UnitaryCircuit)

Create an OnsiteHaar circuit.

LSShadows.xzxcnot_circuit — Method

xzxcnot_circuit(f::UnitaryCircuit)

Create an XZXCnot circuit compossed of Rx-Rz-Rx-Cnot gates with parameter to be dispatched. Here periodic boundary conditions are used

LSShadows.zzx_circuit — Method

zzx_circuit(f::UnitaryCircuit)

Creates a zzx_circuit compossed of Rz-Rzz-Rx gates with parameter to be dispatched. Here periodic boundary conditions are used